requirements.txt
django>=1.9.1
django-ckeditor
django-stdimage
django-templated-email
django-import-export
django-templated-email
django-filter
django-braces
django-tables2
djangorestframework>=3.6.4
django-polymorphic
polymorphic
awesome-slugify
